+++ include/fluxgl/fluxgl.h
@@ -2,3 +2,4 @@
 
 #include "core/window.h"
 #include "core/error.h"
+#include "core/log.h"